core data manager abstraction
