+++ google-news-reader/google-news-reader/API/NetworkManager.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import CoreData
 
 enum NetworkError: ErrorType {
     case NetworkFailure
@@ -70,40 +69,7 @@ class NetworkManager: NSObject {
             if success {
                 
                 // try core data
-                
-                let appDelegate = UIApplication.sharedApplication().delegate as! AppDelegate
-                
-                let managedContext = appDelegate.managedObjectContext
-                
-                var managedArray = [Article]()
-                
-                for articleProto in articleParser.articles {
-                    let managedArticle = Article(prototype: articleProto, inManagedObjectContext: managedContext)
-                    managedArray.append(managedArticle)
-                }
-                
-                // save
-                
-                
-                do {
-                    try managedContext.save()
-                } catch {
-                    print("Error: \(error)")
-                }
-                
-                // try fetch
-                // For debugging only
-//                let fetchRequest = NSFetchRequest(entityName:"Article")
-//                
-//                do {
-//                    let fetchedResults = try managedContext.executeFetchRequest(fetchRequest) as? [NSManagedObject]
-//                    
-//                    print(fetchedResults)
-//                } catch {
-//                    print("Error: \(error)")
-//                }
-                
-                // TODO: If CoreData save fails, throw error or return articles without caching
+                CoreDataManager().cacheFetchedArticles(articleParser.articles)
                 
                 completion(content: articleParser.articles, error: nil)
             } else {
